added methods to animator class

+++ AnimationManager/src/com/jakkarrlgames/AnimationManager/FrameDimension.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.math.Vector2;
 
 public class FrameDimension {
 
-	public float x;
-	public float y;
-	public float width;
-	public float height;
+	public int x;
+	public int y;
+	public int width;
+	public int height;
 }
